# docs/api_doc/swagger.json
{"openapi":"3.0.1","info":{"title":"'@project.name@'","description":"'@project.description@'","contact":{"name":"Joe Doe","email":"dumy@dumy"},"version":"'@project.version@'"},"servers":[{"url":"http://localhost:8080","description":"'@project.description@'"}],"paths":{"/sanction-type/{id}":{"get":{"tags":["sanction-type-controller"],"summary":"Get one fine","description":"Get fine by id","operationId":"getSanctionTypeById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorApi"}}}},"200":{"description":"Successful operation","content":{"*/*":{"schema":{"$ref":"#/components/schemas/SanctionTypeDTO"}}}},"404":{"description":"Sanction type not found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorApi"}}}}}},"put":{"tags":["sanction-type-controller"],"summary":"Update fine state","description":"Update the state of a fine by providing the fine ID and the new state","operationId":"updateSanctionType","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateSanctionTypeDTO"}}},"required":true},"responses":{"200":{"description":"Operation completed successfully","content":{"*/*":{"schema":{"$ref":"#/components/schemas/SanctionTypeDTO"}}}},"404":{"description":"Sanction type not found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorApi"}}}},"500":{"description":"Internal server error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorApi"}}}}}}},"/fine/fine/state":{"put":{"tags":["fine-controller"],"summary":"Update fine state","description":"Update the state of a fine by providing the fine ID and the new state","operationId":"updateFineState","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/FineUpdateStateDTO"}}},"required":true},"responses":{"500":{"description":"Internal server error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorApi"}}}},"200":{"description":"Fine state updated successfully","content":{"*/*":{"schema":{"$ref":"#/components/schemas/FineDTO"}}}},"400":{"description":"Bad Request, invalid input or state transition","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorApi"}}}}}}},"/sanction-type":{"get":{"tags":["sanction-type-controller"],"summary":"Get filtered sanction types","description":"Get filtered sanction types","operationId":"getSanctionTypes","parameters":[{"name":"partialName","in":"query","required":false,"schema":{"type":"string"}}],"responses":{"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorApi"}}}},"200":{"description":"Successful operation","content":{"*/*":{"schema":{"type":"string"}}}}}},"post":{"tags":["sanction-type-controller"],"summary":"Post sanction Type","description":"Post sanction Type","operationId":"createSanctionType","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateSanctionTypeDTO"}}},"required":true},"responses":{"200":{"description":"Operation completed successfully","content":{"*/*":{"schema":{"$ref":"#/components/schemas/SanctionTypeDTO"}}}},"500":{"description":"Internal server error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorApi"}}}}}}},"/proof/claim":{"post":{"tags":["proof-controller"],"summary":"Create a proof for a claim","description":"Creates a new proof for an existing claim by providing details such as proof path, proof type, and claim ID.","operationId":"createProofForClaim","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProofClaimRequestDto"}}},"required":true},"responses":{"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorApi"}}}},"404":{"description":"Claim not found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorApi"}}}},"200":{"description":"Proof for claim created successfully","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ProofResponseDto"}}}},"400":{"description":"Invalid proof request","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorApi"}}}}}}},"/infractions":{"post":{"tags":["infraction-controller"],"summary":"Post  infraction","description":"Post  infraction","operationId":"postInfraction","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateInfractionDto"}}},"required":true},"responses":{"404":{"description":"Plot or Sanction Type Not Found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorApi"}}}},"500":{"description":"Internal server error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorApi"}}}},"200":{"description":"Operation completed successfully","content":{"*/*":{"schema":{"$ref":"#/components/schemas/InfractionDTO"}}}}}}},"/fine":{"post":{"tags":["fine-controller"],"summary":"Post  fine","description":"Post  fine","operationId":"postFine","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateFineDTO"}}},"required":true},"responses":{"200":{"description":"Operation completed successfully","content":{"*/*":{"schema":{"$ref":"#/components/schemas/FineDTO"}}}},"404":{"description":"Plot or Sanction Type Not Found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorApi"}}}},"500":{"description":"Internal server error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorApi"}}}}}}},"/sanction-type/pageable":{"get":{"tags":["sanction-type-controller"],"summary":"Get all sanction types","description":"Find all paginated sanction types o","operationId":"getSanctionTypesPaginated","parameters":[{"name":"page","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":0}},{"name":"size","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":10}},{"name":"partialName","in":"query","required":false,"schema":{"type":"string"}}],"responses":{"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorApi"}}}},"200":{"description":"Successful operation","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Page"}}}}}}},"/ping":{"get":{"tags":["ping-controller"],"summary":"Check healthy of the app","description":"If the app it's alive response pong","operationId":"pong","responses":{"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorApi"}}}},"200":{"description":"Successful operation","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/fine/{id}":{"get":{"tags":["fine-controller"],"summary":"Get one fine","description":"Get fine by id","operationId":"getFineById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"500":{"description":"Internal server error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorApi"}}}},"200":{"description":"Operation completed successfully","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/fine/pageable":{"get":{"tags":["fine-controller"],"summary":"Get all fines","description":"Get all fines","operationId":"getFines","parameters":[{"name":"page","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":0}},{"name":"size","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":10}},{"name":"fineState","in":"query","required":false,"schema":{"type":"array","items":{"type":"string","enum":["ON_ASSEMBLY","APPROVED","REJECTED"]}}},{"name":"sanctionTypes","in":"query","required":false,"schema":{"type":"array","items":{"type":"integer","format":"int64"}}}],"responses":{"500":{"description":"Internal server error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorApi"}}}},"200":{"description":"Operation completed successfully","content":{"*/*":{"schema":{"type":"string"}}}}}}}},"components":{"schemas":{"UpdateSanctionTypeDTO":{"type":"object","properties":{"description":{"type":"string"},"amount":{"type":"number","format":"double"},"user_id":{"type":"integer","format":"int64"},"charge_type":{"type":"string","enum":["PROPORTIONAL","FIXED"]}}},"SanctionTypeDTO":{"type":"object","properties":{"name":{"type":"string"},"description":{"type":"string"},"amount":{"type":"number","format":"double"},"charge_type":{"type":"string","enum":["PROPORTIONAL","FIXED"]}}},"ErrorApi":{"type":"object","properties":{"timestamp":{"type":"string"},"status":{"type":"integer","format":"int32"},"error":{"type":"string"},"message":{"type":"string"}}},"FineUpdateStateDTO":{"required":["fineState","id","updatedBy"],"type":"object","properties":{"id":{"type":"integer","format":"int64"},"fineState":{"type":"string","enum":["ON_ASSEMBLY","APPROVED","REJECTED"]},"updatedBy":{"type":"integer","format":"int32"}}},"FineDTO":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"plotId":{"type":"integer","format":"int64"},"fineState":{"type":"string","enum":["ON_ASSEMBLY","APPROVED","REJECTED"]},"sanctionType":{"$ref":"#/components/schemas/SanctionType"},"infractions":{"type":"array","items":{"$ref":"#/components/schemas/ReducedInfractionDTO"}},"created_date":{"type":"string","format":"date-time"},"created_by":{"type":"integer","format":"int64"},"last_updated_at":{"type":"string","format":"date-time"},"last_updated_by":{"type":"integer","format":"int64"}}},"ReducedInfractionDTO":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"plotId":{"type":"integer","format":"int64"},"description":{"type":"string"},"infractionState":{"type":"string","enum":["APPROVED","ON_APPEALING","CREATED","REJECTED"]},"sanctionType":{"$ref":"#/components/schemas/SanctionType"},"created_date":{"type":"string","format":"date-time"},"created_by":{"type":"integer","format":"int64"},"last_updated_at":{"type":"string","format":"date-time"},"last_updated_by":{"type":"integer","format":"int64"}}},"SanctionType":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"createdDate":{"type":"string","format":"date-time"},"createdBy":{"type":"integer","format":"int64"},"lastUpdatedAt":{"type":"string","format":"date-time"},"lastUpdatedBy":{"type":"integer","format":"int64"},"name":{"type":"string"},"description":{"type":"string"}}},"CreateSanctionTypeDTO":{"type":"object","properties":{"name":{"type":"string"},"description":{"type":"string"},"amount":{"type":"number","format":"double"},"user_id":{"type":"integer","format":"int64"},"charge_type":{"type":"string","enum":["PROPORTIONAL","FIXED"]}}},"ProofClaimRequestDto":{"required":["claim_id","created_by","proof_path","proof_type"],"type":"object","properties":{"proof_path":{"maxLength":255,"minLength":5,"type":"string"},"proof_type":{"type":"string","enum":["TEXT","IMAGE","DOCUMENT"]},"proof_text":{"maxLength":500,"minLength":10,"type":"string"},"claim_id":{"type":"integer","format":"int64"},"created_by":{"type":"integer","format":"int64"}}},"ProofResponseDto":{"type":"object","properties":{"proof_path":{"type":"string"},"proof_type":{"type":"string","enum":["TEXT","IMAGE","DOCUMENT"]},"proof_context":{"type":"string","enum":["CLAIM_JUSTIFICATION","DEFENSE"]},"proof_text":{"type":"string"}}},"CreateInfractionDto":{"type":"object","properties":{"plotId":{"type":"integer","format":"int64"},"description":{"type":"string"},"sanctionTypeId":{"type":"integer","format":"int64"},"claimsId":{"type":"array","items":{"type":"integer","format":"int64"}}}},"ClaimDTO":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"userId":{"type":"integer","format":"int64"},"expiringDate":{"type":"string","format":"date-time"},"sanctionType":{"$ref":"#/components/schemas/SanctionTypeDTO"},"claimState":{"type":"string","enum":["SENT","ON_REVISION","APPEALED","APPROVED","REJECTED"]},"infractionId":{"type":"integer","format":"int64"},"created_date":{"type":"string","format":"date-time"},"created_by":{"type":"integer","format":"int64"},"last_updated_at":{"type":"string","format":"date-time"},"last_updated_by":{"type":"integer","format":"int64"}}},"InfractionDTO":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"plotId":{"type":"integer","format":"int64"},"description":{"type":"string"},"infractionState":{"type":"string","enum":["APPROVED","ON_APPEALING","CREATED","REJECTED"]},"sanctionType":{"$ref":"#/components/schemas/SanctionTypeDTO"},"claims":{"type":"array","items":{"$ref":"#/components/schemas/ClaimDTO"}},"fine":{"$ref":"#/components/schemas/FineDTO"},"created_date":{"type":"string","format":"date-time"},"created_by":{"type":"integer","format":"int64"},"last_updated_at":{"type":"string","format":"date-time"},"last_updated_by":{"type":"integer","format":"int64"}}},"CreateFineDTO":{"type":"object","properties":{"plotId":{"type":"integer","format":"int64"},"sanctionTypeId":{"type":"integer","format":"int64"}}},"Page":{"type":"object","properties":{"totalPages":{"type":"integer","format":"int32"},"totalElements":{"type":"integer","format":"int64"},"size":{"type":"integer","format":"int32"},"content":{"type":"array","items":{"type":"object"}},"number":{"type":"integer","format":"int32"},"sort":{"$ref":"#/components/schemas/SortObject"},"last":{"type":"boolean"},"first":{"type":"boolean"},"pageable":{"$ref":"#/components/schemas/PageableObject"},"numberOfElements":{"type":"integer","format":"int32"},"empty":{"type":"boolean"}}},"PageableObject":{"type":"object","properties":{"offset":{"type":"integer","format":"int64"},"sort":{"$ref":"#/components/schemas/SortObject"},"paged":{"type":"boolean"},"unpaged":{"type":"boolean"},"pageNumber":{"type":"integer","format":"int32"},"pageSize":{"type":"integer","format":"int32"}}},"SortObject":{"type":"object","properties":{"empty":{"type":"boolean"},"sorted":{"type":"boolean"},"unsorted":{"type":"boolean"}}}}}}